Update to improved fyne.Validator per Slack discussion
